+++ CHANGES.rst
@@ -6,8 +6,8 @@ Changelog
 1.1.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
-
+- Fix error message in for loops.
+  [gforcada]
 
 1.1.1 (2018-03-20)
 ------------------